delete on account page

+++ php/goalListDeleter.php
@@ -4,16 +4,16 @@
     $dbPassword = '';
     $guard = false;
     if (isset($_POST['goalID'])) {
-        $goalID=($_POST['goalID']);
+        $goalID=$_POST['goalID'];
     }
     session_start();
+    $userID = $_SESSION["userID"];
     try {
         $db = new PDO($dsn, $dbUsername, $dbPassword);
         $db->setAttribute(PDO::ATTR_ERRMODE, PDO::ERRMODE_EXCEPTION);
-                $query = $db->prepare('DELETE FROM goalslist WHERE goalID = :goalID');
-                $query->bindParam(':goalID', $goalID);
+                $query = $db->prepare("DELETE FROM goalslist WHERE goalID = $goalID AND $userID=userID");
                 $query->execute();
-                header("Location:../html/tracker.php");
+                header("Location:../html/account/accountPage.php");
 
         } catch (PDOException $e) {
         $error_message = $e->getMessage();
